+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatapointop.py
@@ -1287,5 +1287,5 @@ class UpdateDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-22T07:56:11.042Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-22T12:01:38.246Z")
     r"""The date and time the resource was last updated"""